Generate 32-bit FCS even when high bits are zero
The pyDot11.utils.Packet.endSwap() method was adapting the number of bytes
swapped based on the number of bytes needed to represent the integer or
string passed in.  Every place endSwap was used was appending a 32-bit
FCS, so when the high bits of the input number were zero, an incorrect
FCS was computed.

Every place endSwap was called is trying to pack bytes into a frame in
little endian order, so switch the code to use struct.pack and directly
pack to little endian.  Also, delete endSwap so it does not accidentally
get used.

Calling pyDot11.utils.Packet.fcsGen() would centralize the byteswapping
and CRC computation logic and make the code easier to understand.
However, many places fcsGen would be used do not already allocate a
pyDot11.utils.Packet.  Probably it would be best to make fcsGen a free
function or to make it a class method of the pyDot11.utils.Packet class.

Also added a few quick and dirty unit tests for fcsGen.  The tests can be
run by typing `pytest` while in the SRC directory.  The expected values
were computed by running the tests, so might need to be adjusted.

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>